+++ derive/src/newtype.rs
@@ -5,11 +5,12 @@ use syn::Error;
 
 use crate::{
     args::{self, NewTypeName, RenameTarget},
-    utils::{get_crate_name, get_rustdoc, visible_fn, GeneratorResult},
+    utils::{gen_boxed_trait, get_crate_name, get_rustdoc, visible_fn, GeneratorResult},
 };
 
 pub fn generate(newtype_args: &args::NewType) -> GeneratorResult<TokenStream> {
     let crate_name = get_crate_name(newtype_args.internal);
+    let boxed_trait = gen_boxed_trait(&crate_name);
     let ident = &newtype_args.ident;
     let (impl_generics, ty_generics, where_clause) = newtype_args.generics.split_for_impl();
     let inaccessible = newtype_args.inaccessible;
@@ -60,6 +61,7 @@ pub fn generate(newtype_args: &args::NewType) -> GeneratorResult<TokenStream> {
                 inaccessible: #inaccessible,
                 tags: ::std::vec![ #(#tags),* ],
                 specified_by_url: #specified_by_url,
+                directive_invocations: ::std::vec::Vec::new(),
             })
         }
     } else {
@@ -117,6 +119,7 @@ pub fn generate(newtype_args: &args::NewType) -> GeneratorResult<TokenStream> {
         }
 
         #[allow(clippy::all, clippy::pedantic)]
+        #boxed_trait
         impl #impl_generics #crate_name::OutputType for #ident #ty_generics #where_clause {
             fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                 #type_name