+++ derive/src/simple_object.rs
@@ -10,8 +10,8 @@ use crate::{
         self, RenameRuleExt, RenameTarget, Resolvability, SimpleObjectField, TypeDirectiveLocation,
     },
     utils::{
-        gen_deprecation, gen_directive_calls, generate_guards, get_crate_name, get_rustdoc,
-        visible_fn, GeneratorResult,
+        gen_boxed_trait, gen_deprecation, gen_directive_calls, generate_guards, get_crate_name,
+        get_rustdoc, parse_complexity_expr, visible_fn, GeneratorResult,
     },
 };
 
@@ -30,6 +30,7 @@ struct SimpleObjectFieldGenerator<'a> {
 
 pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream> {
     let crate_name = get_crate_name(object_args.internal);
+    let boxed_trait = gen_boxed_trait(&crate_name);
     let ident = &object_args.ident;
     let (impl_generics, ty_generics, where_clause) = object_args.generics.split_for_impl();
     let extends = object_args.extends;
@@ -196,6 +197,18 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
         let visible = visible_fn(&field.visible);
         let directives =
             gen_directive_calls(&field.directives, TypeDirectiveLocation::FieldDefinition);
+
+        let complexity = if let Some(complexity) = &field.complexity {
+            let (_, expr) = parse_complexity_expr(complexity.clone())?;
+            quote! {
+                ::std::option::Option::Some(|__ctx, __variables_definition, __field, child_complexity| {
+                    ::std::result::Result::Ok(#expr)
+                })
+            }
+        } else {
+            quote! { ::std::option::Option::None }
+        };
+
         if !field.flatten {
             schema_fields.push(quote! {
                 fields.insert(::std::borrow::ToOwned::to_owned(#field_name), #crate_name::registry::MetaField {
@@ -213,7 +226,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                     tags: ::std::vec![ #(#tags),* ],
                     override_from: #override_from,
                     visible: #visible,
-                    compute_complexity: ::std::option::Option::None,
+                    compute_complexity: #complexity,
                     directive_invocations: ::std::vec![ #(#directives),* ],
                 });
             });
@@ -377,7 +390,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             }
 
             #[allow(clippy::all, clippy::pedantic)]
-
+            #boxed_trait
             impl #impl_generics #crate_name::resolver_utils::ContainerType for #ident #ty_generics #where_clause {
                 async fn resolve_field(&self, ctx: &#crate_name::Context<'_>) -> #crate_name::ServerResult<::std::option::Option<#crate_name::Value>> {
                     #(#resolvers)*
@@ -387,6 +400,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
             }
 
             #[allow(clippy::all, clippy::pedantic)]
+            #boxed_trait
             impl #impl_generics #crate_name::OutputType for #ident #ty_generics #where_clause {
                 fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                     #gql_typename
@@ -505,6 +519,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
 
             let expanded = quote! {
                 #[allow(clippy::all, clippy::pedantic)]
+                #boxed_trait
                 impl #def_bounds #crate_name::resolver_utils::ContainerType for #concrete_type {
                     async fn resolve_field(&self, ctx: &#crate_name::Context<'_>) -> #crate_name::ServerResult<::std::option::Option<#crate_name::Value>> {
                         #complex_resolver
@@ -513,6 +528,7 @@ pub fn generate(object_args: &args::SimpleObject) -> GeneratorResult<TokenStream
                 }
 
                 #[allow(clippy::all, clippy::pedantic)]
+                #boxed_trait
                 impl #def_bounds #crate_name::OutputType for #concrete_type {
                     fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                         ::std::borrow::Cow::Borrowed(#gql_typename)