+++ derive/src/union.rs
@@ -8,11 +8,14 @@ use syn::{visit::Visit, visit_mut::VisitMut, Error, LifetimeParam, Type};
 
 use crate::{
     args::{self, RenameTarget},
-    utils::{get_crate_name, get_rustdoc, visible_fn, GeneratorResult, RemoveLifetime},
+    utils::{
+        gen_boxed_trait, get_crate_name, get_rustdoc, visible_fn, GeneratorResult, RemoveLifetime,
+    },
 };
 
 pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
     let crate_name = get_crate_name(union_args.internal);
+    let boxed_trait = gen_boxed_trait(&crate_name);
     let ident = &union_args.ident;
     let type_params = union_args.generics.type_params().collect::<Vec<_>>();
     let (impl_generics, ty_generics, where_clause) = union_args.generics.split_for_impl();
@@ -201,7 +204,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
             #(#type_into_impls)*
 
             #[allow(clippy::all, clippy::pedantic)]
-
+            #boxed_trait
             impl #impl_generics #crate_name::resolver_utils::ContainerType for #ident #ty_generics #where_clause {
                 async fn resolve_field(&self, ctx: &#crate_name::Context<'_>) -> #crate_name::ServerResult<::std::option::Option<#crate_name::Value>> {
                     ::std::result::Result::Ok(::std::option::Option::None)
@@ -215,6 +218,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
             }
 
             #[allow(clippy::all, clippy::pedantic)]
+            #boxed_trait
             impl #impl_generics #crate_name::OutputType for #ident #ty_generics #where_clause {
                 fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                     #gql_typename
@@ -242,6 +246,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
                             inaccessible: #inaccessible,
                             tags: ::std::vec![ #(#tags),* ],
                             rust_typename: ::std::option::Option::Some(::std::any::type_name::<Self>()),
+                            directive_invocations: ::std::vec::Vec::new(),
                         }
                     })
                 }
@@ -350,7 +355,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
 
             let expanded = quote! {
                 #[allow(clippy::all, clippy::pedantic)]
-
+                #boxed_trait
                 impl #def_bounds #crate_name::resolver_utils::ContainerType for #concrete_type {
                     async fn resolve_field(&self, ctx: &#crate_name::Context<'_>) -> #crate_name::ServerResult<::std::option::Option<#crate_name::Value>> {
                         ::std::result::Result::Ok(::std::option::Option::None)
@@ -364,6 +369,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
                 }
 
                 #[allow(clippy::all, clippy::pedantic)]
+                #boxed_trait
                 impl #def_bounds #crate_name::OutputType for #concrete_type {
                     fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                         ::std::borrow::Cow::Borrowed(#gql_typename)
@@ -391,6 +397,7 @@ pub fn generate(union_args: &args::Union) -> GeneratorResult<TokenStream> {
                                 inaccessible: #inaccessible,
                                 tags: ::std::vec![ #(#tags),* ],
                                 rust_typename: ::std::option::Option::Some(::std::any::type_name::<Self>()),
+                                directive_invocations: ::std::vec::Vec::new()
                             }
                         })
                     }