+++ src/base.rs
@@ -1,6 +1,7 @@
+#[cfg(not(feature = "boxed-trait"))]
+use std::future::Future;
 use std::{
     borrow::Cow,
-    future::Future,
     sync::{Arc, Weak},
 };
 
@@ -34,7 +35,7 @@ pub trait InputType: Send + Sync + Sized {
     ///
     /// `i32::RawValueType` is `i32`
     /// `Option<i32>::RawValueType` is `i32`.
-    type RawValueType;
+    type RawValueType: ?Sized;
 
     /// Type the name.
     fn type_name() -> Cow<'static, str>;
@@ -64,6 +65,7 @@ pub trait InputType: Send + Sync + Sized {
 }
 
 /// Represents a GraphQL output type.
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 pub trait OutputType: Send + Sync {
     /// Type the name.
     fn type_name() -> Cow<'static, str>;
@@ -85,6 +87,15 @@ pub trait OutputType: Send + Sync {
     fn create_type_info(registry: &mut registry::Registry) -> String;
 
     /// Resolve an output value to `async_graphql::Value`.
+    #[cfg(feature = "boxed-trait")]
+    async fn resolve(
+        &self,
+        ctx: &ContextSelectionSet<'_>,
+        field: &Positioned<Field>,
+    ) -> ServerResult<Value>;
+
+    /// Resolve an output value to `async_graphql::Value`.
+    #[cfg(not(feature = "boxed-trait"))]
     fn resolve(
         &self,
         ctx: &ContextSelectionSet<'_>,
@@ -92,6 +103,7 @@ pub trait OutputType: Send + Sync {
     ) -> impl Future<Output = ServerResult<Value>> + Send;
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: OutputType + ?Sized> OutputType for &T {
     fn type_name() -> Cow<'static, str> {
         T::type_name()
@@ -111,6 +123,7 @@ impl<T: OutputType + ?Sized> OutputType for &T {
     }
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: OutputType + Sync, E: Into<Error> + Send + Sync + Clone> OutputType for Result<T, E> {
     fn type_name() -> Cow<'static, str> {
         T::type_name()
@@ -153,6 +166,7 @@ pub trait InputObjectType: InputType {}
 /// A GraphQL oneof input object.
 pub trait OneofObjectType: InputObjectType {}
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: OutputType + ?Sized> OutputType for Box<T> {
     fn type_name() -> Cow<'static, str> {
         T::type_name()
@@ -162,7 +176,17 @@ impl<T: OutputType + ?Sized> OutputType for Box<T> {
         T::create_type_info(registry)
     }
 
+    #[cfg(feature = "boxed-trait")]
+    async fn resolve(
+        &self,
+        ctx: &ContextSelectionSet<'_>,
+        field: &Positioned<Field>,
+    ) -> ServerResult<Value> {
+        T::resolve(self.as_ref(), ctx, field).await
+    }
+
     #[allow(clippy::trivially_copy_pass_by_ref)]
+    #[cfg(not(feature = "boxed-trait"))]
     fn resolve(
         &self,
         ctx: &ContextSelectionSet<'_>,
@@ -198,6 +222,7 @@ impl<T: InputType> InputType for Box<T> {
     }
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: OutputType + ?Sized> OutputType for Arc<T> {
     fn type_name() -> Cow<'static, str> {
         T::type_name()
@@ -243,6 +268,7 @@ impl<T: InputType> InputType for Arc<T> {
     }
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: OutputType + ?Sized> OutputType for Weak<T> {
     fn type_name() -> Cow<'static, str> {
         <Option<Arc<T>> as OutputType>::type_name()
@@ -262,9 +288,14 @@ impl<T: OutputType + ?Sized> OutputType for Weak<T> {
 }
 
 #[doc(hidden)]
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 pub trait ComplexObject {
     fn fields(registry: &mut registry::Registry) -> Vec<(String, registry::MetaField)>;
 
+    #[cfg(feature = "boxed-trait")]
+    async fn resolve_field(&self, ctx: &Context<'_>) -> ServerResult<Option<Value>>;
+
+    #[cfg(not(feature = "boxed-trait"))]
     fn resolve_field(
         &self,
         ctx: &Context<'_>,