+++ src/resolver_utils/scalar.rs
@@ -61,7 +61,7 @@ pub trait ScalarType: Sized + Send {
 /// #[derive(Serialize, Deserialize)]
 /// struct MyValue {
 ///     a: i32,
-///     b: HashMap<String, i32>,     
+///     b: HashMap<String, i32>,
 /// }
 ///
 /// scalar!(MyValue);
@@ -169,6 +169,7 @@ macro_rules! scalar_internal {
                         inaccessible: false,
                         tags: ::std::default::Default::default(),
                         specified_by_url: $specified_by_url,
+                        directive_invocations: ::std::vec::Vec::new(),
                     }
                 })
             }
@@ -188,6 +189,16 @@ macro_rules! scalar_internal {
             }
         }
 
+        $crate::scalar_internal_output!($ty, $name, $desc, $specified_by_url);
+    };
+}
+
+#[cfg(feature = "boxed-trait")]
+#[macro_export]
+#[doc(hidden)]
+macro_rules! scalar_internal_output {
+    ($ty:ty, $name:expr, $desc:expr, $specified_by_url:expr) => {
+        #[$crate::async_trait::async_trait]
         impl $crate::OutputType for $ty {
             fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
                 ::std::borrow::Cow::Borrowed($name)
@@ -207,6 +218,47 @@ macro_rules! scalar_internal {
                         inaccessible: false,
                         tags: ::std::default::Default::default(),
                         specified_by_url: $specified_by_url,
+                        directive_invocations: ::std::vec::Vec::new(),
+                    }
+                })
+            }
+
+            async fn resolve(
+                &self,
+                _: &$crate::ContextSelectionSet<'_>,
+                _field: &$crate::Positioned<$crate::parser::types::Field>,
+            ) -> $crate::ServerResult<$crate::Value> {
+                ::std::result::Result::Ok($crate::ScalarType::to_value(self))
+            }
+        }
+    };
+}
+
+#[cfg(not(feature = "boxed-trait"))]
+#[macro_export]
+#[doc(hidden)]
+macro_rules! scalar_internal_output {
+    ($ty:ty, $name:expr, $desc:expr, $specified_by_url:expr) => {
+        impl $crate::OutputType for $ty {
+            fn type_name() -> ::std::borrow::Cow<'static, ::std::primitive::str> {
+                ::std::borrow::Cow::Borrowed($name)
+            }
+
+            fn create_type_info(
+                registry: &mut $crate::registry::Registry,
+            ) -> ::std::string::String {
+                registry.create_output_type::<$ty, _>($crate::registry::MetaTypeId::Scalar, |_| {
+                    $crate::registry::MetaType::Scalar {
+                        name: ::std::borrow::ToOwned::to_owned($name),
+                        description: $desc,
+                        is_valid: ::std::option::Option::Some(::std::sync::Arc::new(|value| {
+                            <$ty as $crate::ScalarType>::is_valid(value)
+                        })),
+                        visible: ::std::option::Option::None,
+                        inaccessible: false,
+                        tags: ::std::default::Default::default(),
+                        specified_by_url: $specified_by_url,
+                        directive_invocations: ::std::vec::Vec::new(),
                     }
                 })
             }