+++ src/types/json.rs
@@ -57,6 +57,7 @@ impl<T: DeserializeOwned + Serialize + Send + Sync> InputType for Json<T> {
             inaccessible: false,
             tags: Default::default(),
             specified_by_url: None,
+            directive_invocations: Default::default(),
         })
     }
 
@@ -73,6 +74,7 @@ impl<T: DeserializeOwned + Serialize + Send + Sync> InputType for Json<T> {
     }
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl<T: Serialize + Send + Sync> OutputType for Json<T> {
     fn type_name() -> Cow<'static, str> {
         Cow::Borrowed("JSON")
@@ -87,6 +89,7 @@ impl<T: Serialize + Send + Sync> OutputType for Json<T> {
             inaccessible: false,
             tags: Default::default(),
             specified_by_url: None,
+            directive_invocations: Default::default(),
         })
     }
 
@@ -116,6 +119,7 @@ impl InputType for serde_json::Value {
                 inaccessible: false,
                 tags: Default::default(),
                 specified_by_url: None,
+                directive_invocations: Default::default(),
             }
         })
     }
@@ -133,6 +137,7 @@ impl InputType for serde_json::Value {
     }
 }
 
+#[cfg_attr(feature = "boxed-trait", async_trait::async_trait)]
 impl OutputType for serde_json::Value {
     fn type_name() -> Cow<'static, str> {
         Cow::Borrowed("JSON")
@@ -148,6 +153,7 @@ impl OutputType for serde_json::Value {
                 inaccessible: false,
                 tags: Default::default(),
                 specified_by_url: None,
+                directive_invocations: Default::default(),
             }
         })
     }