+++ src/validation/visitors/cache_control.rs
@@ -9,7 +9,7 @@ pub struct CacheControlCalculate<'a> {
     pub cache_control: &'a mut CacheControl,
 }
 
-impl<'ctx, 'a> Visitor<'ctx> for CacheControlCalculate<'a> {
+impl Visitor<'_> for CacheControlCalculate<'_> {
     fn mode(&self) -> VisitMode {
         VisitMode::Inline
     }